+++ src/common/helpers.ts
@@ -33,6 +33,51 @@ export function calculateRelevanceUsingLength(usedIngredients: SpoonacularIngred
     return toFixNumber(relevance);
 }
 
+export function calculateCheckingAgain(providedIngredients: string[], usedIngredients: SpoonacularIngredient[], missedIngredients: SpoonacularIngredient[]): { relevance: number, missingCount: number } {
+    const mealIngredients = [...usedIngredients.map(ing => ing.name), ...missedIngredients.map(ing => ing.name)];
+
+    return {
+        relevance: calculateRelevance(providedIngredients, mealIngredients),
+        missingCount: calculateMissing(providedIngredients, mealIngredients)
+    };
+}
+
 export function sortDescendingRelevance(meal1: RatedMeal, meal2: RatedMeal): number {
     return meal2.relevance - meal1.relevance;
+}
+
+export function convertAmountToText(amount: number): string {
+    if (amount === 0.33333334) {
+        return '1/3';
+    }
+
+    if (amount === 0.66666667) {
+        return '2/3';
+    }
+
+    if (amount === 0.5) {
+        return '1/2';
+    }
+
+    if (amount === 0.25) {
+        return '1/4';
+    }
+
+    if (amount === 0.5) {
+        return '2/4';
+    }
+
+    if (amount === 0.75) {
+        return '3/4';
+    }
+
+    return amount.toString();
+}
+
+export function compoundTextToTranslate(textAmount: string, unit: string, name: string): string {
+    if (unit === '') {
+        return `${textAmount} ${name}`;
+    }
+
+    return `${textAmount} ${unit} of ${name}`;
 }