+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.51.1] - 2024-09-18
+### Fixed
+- [Marek Kurańda](https://github.com/mjkuranda): Missing ingredient count and relevance index for Spoonacular API.
+- [Marek Kurańda](https://github.com/mjkuranda): Translations of ingredients, changing amount style as well to e.g. `1/4 cup ...`.
+
 ## [1.51.0] - 2024-09-17
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Translating meal description.